Tweaked logos to make them work better in Gaffer.
The main G was made from a red rectangle with a G shaped hole on top of a white rectangle. The white leaked out from underneath when exporting to png from inkscape - replacing it with a white G on a red rectangle fixed that. Also made the greys of the main logo slightly lighter so they don't blend into Gaffer's default grey background colour.