Added exposure and gamma controls to the image viewer.
This is achieved by setting a preprocessor containing a grade node with appropriate settings. This does mean that any subclasses which were formerly calling setPreprocessor() to perform a conversion will have to be modified to take advantage of the new behaviour - see ImageViewTest.testDeriving for an example.

The viewer has an outstanding bug (which already had a todo) whereby the sampled colour at the mouse x,y coordinates includes the additional transforms, whereas it should not. I haven't addressed this here, but instead have added todo items describing the approach I think we should take.

Fixes #571.