Fixed GadgetWidget overlay event handling problems.
Clicking in the main part of the GadgetWidget wasn't taking focus back from the overlay, and typing in the GadgetWidget wasn't delivering events properly to overlays with the focus. We also weren't passing mouse move on to the overlay all the time, and since the overlay needs it to do proper enter/move/leave event handling, things like highlighting of hovered buttons were broken.