Optimised OpenColorIO with a new ColorProcessor base class.
Because the OCIO transform might mix the R, G, and B channels, the output for any channel depends on all three inputs. We were dealing with this very naively by doing all the work again for each channel. The ColorProcessor base class makes it easy to do this work just once, caching the results on the intermediate colorDataPlug() - the computeChannelData() implementation is therefore dead simple, simple passing out the right component from the intermediate results.

The new ColorProcessor base class also introduces a couple of conventions we should use everywhere (and backport to existing classes in time) :

- ColorProcessor::affectsColorData(). This forms a matching trio with hashColorData() and processColorData(), meaning that a derived class can do all of its work without having to override any of the virtual functions in more distant ancestors (previously it would have been necessary to override affects()).
- Interleaving the hash*() and matching compute*() methods in the .cpp file. Keeping the each hash next to its corresponding compute makes it much easier to verify that they are consistent.