Fixed Grade channel-independence hash bugs.
All three components of the colour plugs that control the grade were being hashed in for each channel, when in reality only a single component affects each channel. Fixing this should result in faster interaction and smaller caches when the user is tweaking a single channel at a time.

Fixes #338.