Made Grade::channelEnabled() disabled for default plug values.
This means that it is a pass-through for the most common case in the image viewer, yielding improved performance. Also now accessing only the components of the input plugs that are needed for each channel, rather than getting the colour and then extracting the component - this also has a small performance improvement.