Fixed ImagePrimitiveSource hashing bug.
This was somewhat hard to track down, in that it only appeared when an ImagePrimitiveSource derived node was connected to a disabled node. The ImagePrimitiveSource was failing to hash in the tile origin and channel name, but this was going unnoticed because it turns off caching for its output plug, so the hash wasn't used to do cache lookups which would show the problem. When it was connected to a disabled node, the hash was passed through the disabled node, but the data was read via a pass-through compute() on the disabled node, on a plug which has caching enabled. That caching then fell foul of the hash bug. Just putting the tile origin and channel name fixes the hash and therefore the bug.

This does raise an interesting question though, in that we don't really want caching to start happening just because of a pass-through, but that is a larger question for the future...

Fixes #420.