Fixed SceneReader hashing, and made childNames hash constant.
The SceneReader::hash() method was lazily hashing in time for everything, regardless of whether it was even related to outPlug(). It now correctly implements the hash*() methods instead, but omits hashChildNames() because the hierarchy provided by SceneInterfaces is static over time. This shaves approximately 6% off the time to traverse 20 frames of an animated scene.