Refactored Timeline to use new Playback class.
The Playback class provides a central controller for all clients who wish to initiate playback or react when playback occurs. Playback classes exist for each Context so that different Editors can have different playback states if they also have different contexts. Signals on the Playback class can be used to be notified when the playback frame range changes or when playback starts or stops.

This also fixes the following bugs in the Timeline class :

- Wraparound during playback would remove the fractional part of the frame number.
- Entering a value in the frame field which was outside the frame range would make it impossible to subsequently enter a fractional frame number.
- Changing the frame number via the numeric field didn't update until focus left the field.