Optimised SceneHierarchy update for improved playback performance.
When I say optimised, I mean that SceneHierarchy update is now disabled completely during playback. It's incredibly unlikely that the hierarchy changes with time (although it is allowed), and during playback the user is far more interested in the contents of the Viewer than of the SceneHierarchy, so wasting time updating it can't be justified. A final update it performed when playback stopped, in case the hierarchy has changed in some way.

This change alone reduces the time to playback a 20 frame test animation by 75%. The improvement should be even greater in conjunction with the SceneReader optimisations currently being reviewed, because once they are merged, SceneHierarchy update would represent an even greater proportion of playback processing.