Added 'auto expand' mode to the scene viewer.
This is implemented with a "minimumExpansionDepth" plug on the SceneView class, which specifies a number of levels to always expand by, regardless of any other settings. Currently this is exposed simply as an on/off menu item in the viewer toolbar as Ben originally requested it. Also added menu items for the existing expand/collapse selection functionality, which should make the keyboard shortcuts a bit more discoverable even if the menu items aren't that convenient in themselves.

This continues the accumulation of nonsense into the SceneProcedural in support of viewer functionality, so I added a todo to remove all the nonsense when we find the time to have the viewer do its own conversion of the scene to OpenGL.

Fixes #163