Added Shift+Down shortcut for full expansion in the scene viewer.
The logic for this is a little complicated, and the chances are high that we'll refactor this code in the future when the viewer does its own scene management, so I also added test cases to check that things work as expected in the new code and will continue to in the future. This also fixes one bug, where leaf level items were being marked as expanded, meaning that they'd require two presses of the Up key to get them to collapse.

Fixes #556.