Added array data support to PlugValueWidget drag and drop.
If an array of length 1 is dragged onto a PlugValueWidget, and the array element if of an appropriate type, then the widget will accept the drag. The biggest benefit of this is that objects can now be dragged from the Viewer and SceneHierarchy into path fields such as the AimConstraint target or the StandardOptions camera.