Implemented "look through camera" in SceneView.
This is a first pass, and will need updating to support properly showing a resolution gate with appropriate padding. That gets into much bigger questions of how we represent cameras in Cortex, how we convert them out of Maya and Houdini and how we create them in Gaffer though, so will need further discussion.