Moved Clamp::channelEnabled() shortcuts to Clamp::enabled().
Because the shortcut condition isn't channel dependent, it's slightly more correct to implement it at the enabled() level rather than the channelEnabled() level. It also removes a tiny amount of processing, although I doubt that is significant.

Also fixed the test case for checking pass-through hashes - it was testing the hash of the image data after retrieving it. This merely checks that the node hasn't changed the data during compute, but this alone is not enough to ensure that redundant computation is not performed and that cache entries are shared between a disabled Clamp node and its input. For that to happen, the result of the Clamp::hash*() methods must be a pass-through of the input hash*() methods, which is what the test now checks.