Removed nasty OIIO versioning workaround.
We were renaming the standard build of the OpenImageIO library so that it didn't conflict with the one shipped with Arnold. Arnold now (in 4.1) links OpenImageIO statically and hides the symbols, so we no longer need the workaround.

Also tweaked the OSL build to install all the OSL binaries including testshade and testrender. These still aren't shipped in the final gaffer package, but it's useful to have them there during development for testing purposes.