Added ImageTimeWarp, ImageContextVariables and ImageSwitch.
These are implemented using the same mechanism as SceneTimeWarp, SceneContextVariables and SceneSwitch - simply mixing in functionality from the generic classes in the Gaffer module.

Deprecated GafferImage::Select as it is now superceded by ImageSwitch - it remains in the build for backwards compatibility but has been removed from the node menu.