Added RecursionPredicate to FilteredRecursiveChildIterator.
The existing Predicate dictated what type of GraphComponents would be visited by the iterator, but provided no means of specifying when recursion should happen - so the iterator would traverse the entire tree looking for items matching the predicate.

The new RecursionPredicate controls the pruning of the recursion, providing an easy means of avoiding recursing into parts of the hierarchy within the iterator itself, without having to manually prune all the time. The default value for this predicate keeps compatibility with existing code.