Redefined RecursivePlugIterators to not recurse to other nodes.
Prior to this, when iterating the plugs on one node, the iteration would also recurse to the plugs of any child nodes of the root node. This is not what we want generally, and not what our current use cases of the plug iterators assume. This fixes the recently added nested Boxes bug, and no doubt a number of other problems which were biding their time before showing themselves.

It's still possible to perform the previous sort of recursion using a custom FilteredRecursiveChildIterator typedef if necessary, but no typedefs are provided for it out of the box because it seems to not be a common pattern.