Added enabling/disabling to ContextProcessor.
The ContextProcessor was written before the DependencyNode enabledPlug() and correspondingInput() mechanism was defined, so it didn't implement it at all. It did however inherit enabled plugs from SceneProcessor and ImageProcessor which made it look like it should. Now it correctly implements them, passing through a perfect (cache sharing) copy of the input when disabled.

This provides proper enabling/disabling for SceneTimeWarp, ImageTimeWarp, SceneContextVariables and ImageContextVariables.