Defined supported closures directly in OSLRenderer.
In future OSL versions (the current master which will presumably become 1.5), there are no built in closures defined by the ShadingSystem, so we must define them ourselves. For current versions, this has the benefit of limiting the available closures to only emission, so that we no longer advertise closures that make no sense in our no-illumination context.