Fixed PySide incompatibility in Slider.
PyQt supports the passing of points directly to drawPolygon(), and PySide supports passing them as a list. Neither support the other's syntax. Fortunately they both support the creation of a QPolygonF from a list and then accept that in drawPolygon(). This problem manifested itself when drawing the off-the-end indicator for sliders.