Fixed colour space used for ImageView colour sampling.
It was sampling colours after gamma, exposure and luts were applied. Now it samples them before while the image is in linear space. This is a first step in refactoring the ImageView so that the image stats and under-the-mouse samples are computed by nodes and displayed by PlugValueWidgets in an overlay - all the computation is in nodes now but the display is still using miscellaneous OpenGL calls.