Fixed copy/paste bug with unselected PathFilter input.
When the copy is taken, the output of the PathFilter is calculated to provide a static value for the filter input on the node being copied. Because the usual ScriptNode context doesn't contain a scene:path entry, this was erroring. We now ignore missing scene:path entries, and just return NoMatch instead.