Added space conversion methods to GafferImage::Format.
Because Gaffer uses a Y-up coordinate system and Cortex and OpenEXR use Y-down, it's handy to have this documented somewhere and to have some methods for easily converting between coordinate systems.

Also added a todo in ImagePrimitiveSource.inl, to fix the display window computation and use these new convenience methods to make the data window and channel data match appropriately.