Improved GafferImage::Display node hashing.
Prior to this, the hash for every tile in the image was changed any time a new bucket arrived. This meant a lot of churning in the cache and a lot of unnecessary computation in the viewer, as exposure, gamma and lut were repeatedly being reapplied to tiles which hadn't changed but had new hashes.

This improves viewer responsiveness approximately 2x during an interactive render.