Optimised ImageView internal connections.
Removing connections and replacing them with a plugSet callback which transfers the values gives a rather disturbing 1.5x improvement in ImageView update when viewing a Display node. This implies that there are some optimisations to be made when propagating static values through plug connections (we're currently relying on an unnecessary fake computation to do that every time the value is accessed, whereas we could be doing it as a one off operation before compute), but in the meantime it seems worth making this simple change which greatly improves the use of the render viewer.