Fixed GafferImageTest.DisplayTest.
It was relying on waiting a fixed amount of time for data to be received and processed on the DisplayDriverServer thread - sometimes the processing would take longer and the test would be performed before the data was processed, yielding failure. We now wait on semaphores connected to Display.dataReceivedSignal() and Display.imageReceivedSignal() to make sure the tests are performed at the appropriate time.