Implemented per-instance metadata for Box nodes.
Since each Box may be created for a different purpose, and have different plugs to any other Box, it doesn't make sense for them to have static metadata entries. The Box::setPlugMetadata() method therefore allows metadata to be set on a per-plug-per-node basis. This will provide the basis for the user to make further customisations to Boxes, like setting descriptions for plugs and so on.