Simplified OpDialogue error display.
We were putting the entire error message inside a label, which meant that the window could become extremely large if the error was long. We now simply display a short "Failed" message and rely on the Details pane to provide the full information - this is the same thing we do for successful operation - there is a short "Completed" message and the result is displayed in the details pane.