Refactored BranchCreators to avoid unnecessary hashing overhead.
There were various circumstances where we were calling the base class hash methods only to reassign to the hash rather than append to it. By moving the responsibility of calling the base class from BranchCreator to its subclasses, we can avoid that overhead. The only results in a very mild (a few percent) speedup in my test case, but it also brings the branch hashing logic in line with that used elsewhere, which hopefully helps make everything a little more understandable.