Removed BranchCreator "name" plug.
It should be entirely up to the derived classes what the created branch looks like, and it should be possible to have multiple root entries in the branch - this is to facilitate the creation of a Parent node. Instancer and Seeds nodes still have a "name" plug, but are now responsible for using it themselves.