Fixed BranchCreator to preserve existing sibling branches.
Before, any existing branches below the parent were destroyed and replaced by the new branch - now they are kept and the new branch is renamed to avoid clashes if necessary. Also added a \todo with some thoughts about how we might be able to generalise this mechanism in the future.