Fixed bug in read only Plugs with input connections.
They would throw when the value was being transferred through the connection. Readonlyness now only prevents the setting of static values on plugs, and has no bearing on setValue() calls made during computation or while propagating values through connections. This is reasonable because in the latter cases, no modification is made to the plug - the values are all stored in the cache instead.