Fixed some "invalid drawable" error messages on OS X.
I had hoped that this fix would make the previous commit unnecessary, but testing shows that they address different problems and both are needed. Additionally, although this fixes the initial "invalid drawable" messages on startup, they still appear later on when switching between fullscreen and not-fullscreen. This goes some way to addressing #396, but I suspect it'll be very hard to fix it completely without patching Qt in some way.