Fixed StandardNodeToolbar for dynamic addition and removal of plugs.
This is using the same method as is used in StandardNodeUI and CompoundPlugValueWidget. The method itself has flaws (although they're much more likely to be an issue for CompoundPlugValueWidget than for the toolbars), so I've added some more notes as to how we might be able to address them.

Also added workaround for labels having an unwanted width, and a todo for getting it fixed.