Enabled input connections to PathFilter "paths" plug.
This allows it to be promoted to box level, be driven by expressions etc. It's feasible for a very nefarious user to drive it with an expression which varies based on the "scene:path" context entry, in which case descendant and ancestor queries would be inaccurate, but a user that nefarious deserves what they get.

Fixes #704.