Added pre-selection highlighting for GraphGadget.
As the user drags out the selection rectangle in the node graph, nodes are highlighted to indicate whether or not they will be selected. This makes it clearer what is about to happen, and avoids ambiguity about whether it is wholly enclosed or merely intersected nodes that will be selected.

Fixes #94.