Fixed StandardGraphLayout test failures.
These were brought about by 812fafc4e4f32cda50362185350494851c76afec, because it meant that connections were being made to the "enabled" plug rather than the expected one. Disabling the creation of nodules for the "enabled" plug fixes the tests, and makes the behaviour of AddNode match all other nodes with "enabled" plugs, whereby that plug is not shown in the graph.