Refactored ConnectionGadget to allow for custom subclasses.
ConnectionGadget itself is now an abstract base class with a factory mechanism for the registration and creation of subclasses. StandardNodeGadget takes on most of the old functionality of the ConnectionGadget, matching the pattern we have for NodeGadget and Nodule.

At its simplest, the factory mechanism allows the customisation of styles and tooltips as illustrated in the test cases, but it would now be possible to implement entire new derived classes for totally custom drawing of connections.