Added graph layout algorithm and Edit/Arrange menu item.
The algorithm is invoked via a new pure virtual GraphLayout::layoutNodes() method, which is implemented in the StandardGraphLayout class. The StandardGraphLayout uses a new internal LayoutEngine class which converts the GraphGadget contents into boost::graph form, and then uses a hybrid combination of force directed layout and hard constraints to produce something resembling a sensible layout, before transferring it back to the GraphGadget.

The hybrid approach was necessary because existing layered graph drawing algorithms don't apply too well to our graphs where connections flow both left->right /and/ top->bottom. Pleasingly, it also can be seen as a generalisation of the existing system we had for positioning nodes one at a time in positionNode() or as a group in positionNodes(), where the former soft constraints are treated as springs and iteration over the hard constraints is used to allow more complex constraint relationships to be resolved, along with collisions. This means that positionNode() and positionNodes() are now implemented using the same LayoutEngine as layoutNodes().

Hopefully fixes #638 sufficiently to be given a go in the wild before we make new issues for further refinement.