Updated GraphGadget for new layout code.
Removed inappropriate call to positionNode() in GraphGadget. There's a significant overhead to this now, and it's also unnecessary, because the creator of the node is responsible for calling one of the layout functions at the appropriate time. Removed \todo which suggested that GraphGadget should automatically figure out the right thing to do get new nodes laid out properly - I now think that this requires external knowledge unavailable to the GraphGadget (e.g. is this an import operation, or a paste, or a new node, or a whole load of scripted operations).