Added failing NodeGraph test case.
The problem is that GraphGadget makes a position plug immediately upon a new node being added, in addNodeGadget(). But when a node with an existing position plug is being pasted, it will add a second position plug immediately after, resulting it being named and becoming a useless duplicate. Perhaps ideally we shouldn't add the plug at all in addNodeGadget(), but we currently rely on it existing in offsetNodes(), so it's not totally clear what the best approach is.