Fixed disappearing 0 values in OSLObject and OSLImage.
The problem was that we were using the weight of a debug closure to specify the value for channels and primitive variables. OSL optimises away any closures with a zero weight, so we were ending up with no closure, and therefore were unable to correctly manipulate the image/object. Storing the value as an internal attribute on the debug closure and always using a weight of 1 works around this issue.