Replaced OSL Sum with OutObject and OutImage shaders.
Although these have identical functionality to the Sum shader (they just sum a bunch of closures), it's good to have shaders with names that make it clearer what their purpose is. Later on we may also be able to use metadata to limit the input closures to only come from the appropriate Out* shaders.