Fixed unwanted copy/paste Plug.setValue() calls.
When the input to a plug wasn't part of the selection being copied, we were computing the input value and storing it as a setValue() call in the serialisation. This was bogus, because any computed value can be dependent on any part of a Context, and the Context in which we perform the serialisation may not be suitable for computing the value. Now copy/paste with excluded connections causes the plug to revert to its default value, which is more like the disconnection behaviour.

There is an argument that we should actually be outputting a setValue() call, but for the value that would exist if the connection was removed, but for now we've decided that this approach is sufficient to fix #740.

As for a fuller fix, there are still some open questions about how disconnection should affect the plug anyway. Perhaps for plugs with inputs which aren't from a ComputeNode (just a direct connection from a promoted plug for instance), there is an argument that they should retain the connected value rather than revert to their previous value. When we resolve these questions we could also revisit this issue.