Refactored FormatPlugSerialiser to subclass ValuePlugSerialiser.
This allows us to remove a bunch of code which appears just to have been copy pasted a long time ago, since when we have made changes to the original. Also removed comment which appeared not to reflect the code at all, and added a todo to find out why we need to serialise all the registerFormat() stuff anyway.