Fixed serialisation of Constant node with zero alpha.
Rather than correctly set the default value for the color plug in the constructor, the Constant node was constructing with a default of 0, and then in the constructor calling setValue( 1 ). This breaks completely with the default-value-omission optimisation, because when the user sets the value back to the true default, the value isn't saved out, and then on reloading the constructor clobbers the true default.