Made CompoundDataPlug::addMember() set plug default values.
This is necessary to avoid the same serialisation-of-default-values problem as was addressed in the Constant node in the previous commit. It also improves things from a user perspective - setting a value in an Attributes or Options node back to default now reverts it to the value it had on construction rather than always setting it to 0 or "".