Removed unreasonable todo item from ComputeNodeTest.
This dates back to a time before contexts were introduced and there was effectively a single global context. Back then each plug actually had a dirty flag, representing whether or not it had been computed for that context yet. That is no longer the case, and the idea of a flag defeats the object of multiple contexts and theaded evaluation.