Fixed test cases for wrapping changes.
Prior to this, a lot of unnecessary wrapping and garbage collection for wrapped-but-not-subclassed objects meant that we could use "is", but now we must use ".isSame()" to work around object identity problems. That's the price we pay for improved performance in this case. Note that in general ".isSame()" has always been necessary for RefCounted types, and we were just lucky that the particular types used in those tests had been wrapped before.