Fixed crash in GraphGadget destruction.
This one is quite interesting. The GraphGadget stores its root node with a smart pointer in m_root. If noone else has a reference to the root, then the root node is deleted during the destruction of the GraphGadget. Prior to our wrapper changes this wouldn't occur immediately because the ScriptNode would require garbage collection by IECore::WrapperGarbageCollector - this essentially delayed actual destruction of the root until well after the GraphGadget was destructed fully.

Now the root is deleted immediately. While it is doing this, it removes all its children, which in turn remove all their connections, which in turn signal to the GraphGadget that something has changed. But the GraphGadget is destructing, so it blows up trying to deal with the signal. But surely all the signals should be disconnected from the GraphGadget automatically on destruction, because it derives from boost::signals::trackable? Well sure, but that will only happen when the trackable base class destructor is called, which is /after/ GraphGadget has destructed - so too late. The solution is to use scoped_connections to store the connection. These are stored in m_nodeGadgets, which is guaranteed to be destructed before m_root (members destruct in reverse order of initialisation). This ensures GraphGadget is disconnected from all signals before m_root dies, and GraphGadget can die in peace.