Added workaround for crash in InteractiveRenderManRenderTest.
The comments describe the situation whereby the InteractiveRenderManRender node was evaluating the scene while the connections were in an invalid intermediate state. This was actually occurring during the destructor for the ScriptNode containing the renderer, which removes all children (and therefore their connections). That destructor is now triggered much earlier because it no longer requires garbage collection, unmasking a problem which has existed all along.

The comment also describes a proper fix for the problem which we should implement when time allows.