Reimplemented StandardNodeToolBar with new GafferUI.PlugLayout class.
The old internal implementation was vulnerable to plugs changing names, as documented in the comments. The new PlugLayout class doesn't have this vulnerability as it uses the fact that IECore.Refcounted __hash__ and __eq__ are now implemented sufficiently to use a RefCounted object as a dictionary key. The same vulnerability exists in the implementations for CompoundPlugValueWidget and StandardNodeUI - the intention is to refactor both of these over time to use the PlugLayout class, and improve the PlugLayout to allow sectioning driven by Metadata, allowing it to provide user-designed UIs for the Box class.

It's necessary to introduce a new class here rather than improve on CompoundPlugValueWidget for a couple of reasons :

- CompoundPlugValueWidget can only display plugs which are children of other plugs - as it derives from PlugValueWidget it must accept only parents which are themselves Plugs. This makes it unsuitable for use in the top level uis in StandardNodeUI, where the plugs are held directly on a Node.
- CompoundPlugValueWidget has become quite unwieldy with many subclasses implementing various _protected() methods that customise its behaviour. Over time we should replace the CompoundPlugValueWidget internals with PlugLayout too, but that will be a bigger undertaking whereby we'll have to deprecate various things and rejig several derived classes. In the process of doing this we need to reduce the coupling between all these components, using Metadata as much as possible to control behaviour instead.