Updating current selection in PathListWidget.setSelectedPaths().
The QTreeView supports a set of selected (highlighted) items, and also a "current" item, which is not necessarily selected but is the focus of keyboard operations and the like. The call to selectionModel.clear() clears both of these, so it's good to place the "current" item back at the top of the selection so that keyboard operations start where the user would expect.