Fixed CompoundPlug plugSetSignal() emission when children change.
Previously we only emitted plugSetSignal for the CompoundPlug itself, and not all its ancestors up to the node as was done for normal plugSetSignal emission. Now the same code path is used in both scenarios.

This problem came to light when the ScriptNode failed to update its custom context variables when plugs were added during loading. This then became obvious when any bookmarks referring to those variables failed to evaluate.