Fixed Reference plug duplication.
Plugs which were loaded from a reference, but had no connections within the reference were being saved as dynamic plugs on the reference node, meaning that they doubled up next time the script was loaded. Reference::isReferencePlug() is now implemented to match the plug filtering in  Box::exportForReference().

Note that this does mean that any plugs which were added to reference nodes directly without an "__" prefix will no longer be saved, as reflected in the changes to ReferenceTest.testReloadDoesntRemoveCustomPlugs(). This is worth careful consideration during review - @davidminor could you take a look at this in relation to the IE internal code?

This change is an interim measure until we have the opportunity to break compatibility and implement the ideas in #801.