Fixed overzealous Viewport drag tracking.
We would track drags whenever the mouse was in the autoscroll region at the edge of the viewport - even if the drag started there and the user was dragging the other way. This fix ensures that the drag doesn't start until the user has dragged further out of the viewport than their start position.

Fixes #550.